MAJOR UPDATE (see details)
Updates:
1-Virtual Memory implemented (MMU)
2- Special control instructions implemented (MSR, MRS, LIVP, SIVP, LEVP, etc...)
3- Interrupts, Exceptions, CPU Modes, Permissions and Fault mechanisms implemented
4- Util libraries' API updated (Cmdline)
5- Code cleanup and program header/banner updated (with cool ASCII art)

Note: We're giving up on the FPU. It'll stay on the design documents though.

TODO:
1- Take care of banked registers
2- DEBUG!!!!!!!!!!!!!

_____________________
Note to self:
The only way we can properly debug anything from now on is to use the LLVM C Compiler and ditch the old assembler. I'll have to find a way to link multiple ELF objects together using clang's linker and/or to relocate/link them with a small utility program of my own (written in C++ obviously). I also need to prepare an ELF file to be properly loaded to address 0 in memory ready to be booted up. Only then i can properly test the more complex features of the CPU effectively, such as the Virtual Memory.